+++ crates/holochain/src/conductor/conductor.rs
@@ -125,6 +125,9 @@ pub use builder::*;
 mod chc;
 pub use chc::*;
 
+mod conductor_services;
+pub use conductor_services::*;
+
 pub use accessor_impls::*;
 pub use app_impls::*;
 pub use app_status_impls::*;
@@ -228,6 +231,8 @@ pub struct Conductor {
     post_commit: tokio::sync::mpsc::Sender<PostCommitArgs>,
 
     scheduler: Arc<parking_lot::Mutex<Option<tokio::task::JoinHandle<()>>>>,
+
+    pub(crate) services: RwShare<Option<ConductorServices>>,
 }
 
 impl Conductor {
@@ -239,6 +244,8 @@ impl Conductor {
 
 /// Methods related to conductor startup/shutdown
 mod startup_shutdown_impls {
+    use kitsune_p2p_types::box_fut_plain;
+
     use crate::conductor::manager::{spawn_task_outcome_handler, OutcomeReceiver, OutcomeSender};
 
     use super::*;
@@ -272,6 +279,7 @@ mod startup_shutdown_impls {
                 keystore,
                 holochain_p2p,
                 post_commit,
+                services: RwShare::new(None),
             }
         }
 
@@ -325,6 +333,21 @@ mod startup_shutdown_impls {
                 }
                 let task = spawn_task_outcome_handler(self.clone(), outcome_rx);
                 *lock = Some(task);
+            });
+
+            self.services.share_mut(|services| {
+                let mut dpki = MockDpkiService::new();
+                dpki.expect_is_key_valid()
+                    .returning(|_, _| box_fut_plain(Ok(true)));
+                dpki.expect_key_mutation()
+                    .returning(|_, _| box_fut_plain(Ok(())));
+
+                let app_store = MockAppStoreService::new();
+
+                *services = Some(ConductorServices {
+                    dpki: Arc::new(dpki),
+                    app_store: Arc::new(app_store),
+                });
             });
 
             self.clone().add_admin_interfaces(admin_configs).await?;
@@ -628,8 +651,6 @@ mod dna_impls {
                     .filter_map(|cell_id| cells.remove(cell_id).map(|c| (cell_id, c)))
                     .collect()
             });
-            self.running_cells.share_ref(|cells| dbg!(cells.len()));
-
             for (cell_id, item) in to_cleanup {
                 if let Err(err) = item.cell.cleanup().await {
                     tracing::error!("Error cleaning up Cell: {:?}\nCellId: {}", err, cell_id);
@@ -732,7 +753,10 @@ mod network_impls {
     use holochain_conductor_api::NetworkInfo;
     use holochain_p2p::HolochainP2pSender;
     use holochain_zome_types::block::Block;
-    use holochain_zome_types::block::BlockTargetId;
+
+    use crate::conductor::api::error::{
+        zome_call_response_to_conductor_api_result, ConductorApiError,
+    };
 
     use super::*;
 
@@ -781,17 +805,17 @@ mod network_impls {
 
         /// Block some target.
         pub async fn block(&self, input: Block) -> DatabaseResult<()> {
-            Ok(self.spaces.block(input).await?)
+            self.spaces.block(input).await
         }
 
         /// Unblock some target.
         pub async fn unblock(&self, input: Block) -> DatabaseResult<()> {
-            Ok(self.spaces.unblock(input).await?)
+            self.spaces.unblock(input).await
         }
 
         /// Check if some target is blocked.
         pub async fn is_blocked(&self, input: BlockTargetId, timestamp: Timestamp) -> StateQueryResult<bool> {
-            Ok(self.spaces.is_blocked(input, timestamp).await?)
+            self.spaces.is_blocked(input, timestamp).await
         }
 
         pub(crate) async fn prune_p2p_agents_db(&self) -> ConductorResult<()> {
@@ -1035,6 +1059,45 @@ mod network_impls {
             debug!(cell_id = ?call.cell_id);
             let cell = self.cell_by_id(&call.cell_id)?;
             Ok(cell.call_zome(call, Some(workspace_lock)).await?)
+        }
+
+        /// Make a zome call with deserialization and some error unwrapping built in
+        pub async fn easy_call_zome<I, O, Z, F>(
+            &self,
+            provenance: &AgentPubKey,
+            cap_secret: Option<CapSecret>,
+            cell_id: CellId,
+            zome_name: Z,
+            fn_name: F,
+            payload: I,
+        ) -> ConductorApiResult<O>
+        where
+            FunctionName: From<F>,
+            ZomeName: From<Z>,
+            I: Serialize + std::fmt::Debug,
+            O: serde::de::DeserializeOwned + std::fmt::Debug,
+        {
+            let payload = ExternIO::encode(payload).expect("Couldn't serialize payload");
+            let now = Timestamp::now();
+            let (nonce, expires_at) = holochain_state::nonce::fresh_nonce(now)?;
+            let call_unsigned = ZomeCallUnsigned {
+                cell_id,
+                zome_name: zome_name.into(),
+                fn_name: fn_name.into(),
+                cap_secret,
+                provenance: provenance.clone(),
+                payload,
+                nonce,
+                expires_at,
+            };
+            let call =
+                ZomeCall::try_from_unsigned_zome_call(self.keystore(), call_unsigned).await?;
+            let response = self.call_zome(call).await;
+            match response {
+                Ok(Ok(response)) => Ok(zome_call_response_to_conductor_api_result(response)?),
+                Ok(Err(error)) => Err(ConductorApiError::Other(Box::new(error))),
+                Err(error) => Err(error),
+            }
         }
     }
 }
@@ -1351,7 +1414,7 @@ mod clone_cell_impls {
                         .join(", ");
 
                     ConductorError::CloneCellError(format!(
-                        "no base cell found for provided role id. Available role names are: ({})",
+                        "no base cell found for provided role name. Available role names are: ({})",
                         role_names
                     ))
                 })?;
@@ -2429,12 +2492,26 @@ impl Conductor {
         // add clone cell to app and instantiate resulting clone cell
         let (_, installed_clone_cell) = self
             .update_state_prime(move |mut state| {
+                let state_copy = state.clone();
                 let app = state.get_app_mut(&app_id)?;
                 let agent_key = app.role(&role_name)?.agent_key().to_owned();
-                let cell_id = CellId::new(clone_dna_hash, agent_key);
-                let clone_id = app.add_clone(&role_name, &cell_id)?;
+                let clone_cell_id = CellId::new(clone_dna_hash, agent_key);
+
+                // if cell id of new clone cell already exists, reject as duplicate
+                if state_copy
+                    .installed_apps()
+                    .iter()
+                    .flat_map(|(_, app)| app.all_cells())
+                    .any(|cell_id| *cell_id == clone_cell_id)
+                {
+                    return Err(ConductorError::AppError(AppError::DuplicateCellId(
+                        clone_cell_id,
+                    )));
+                }
+
+                let clone_id = app.add_clone(&role_name, &clone_cell_id)?;
                 let installed_clone_cell = ClonedCell {
-                    cell_id,
+                    cell_id: clone_cell_id,
                     clone_id,
                     original_dna_hash,
                     dna_modifiers,