+++ crates/holochain/src/conductor/space.rs
@@ -28,6 +28,7 @@ use holochain_sqlite::{
     prelude::{DatabaseError, DatabaseResult},
 };
 use holochain_state::{
+    host_fn_workspace::SourceChainWorkspace,
     mutations,
     prelude::{from_blob, StateQueryResult},
     query::{map_sql_dht_op_common, StateQueryError},
@@ -37,7 +38,7 @@ use holochain_types::{
     db_cache::DhtDbQueryCache,
     dht_op::{DhtOp, DhtOpType},
 };
-use holochain_zome_types::{Entry, EntryVisibility, SignedAction, Timestamp};
+use holochain_zome_types::{DnaDef, Entry, EntryVisibility, SignedAction, Timestamp};
 use kitsune_p2p::{
     event::{TimeWindow, TimeWindowInclusive},
     KitsuneP2pConfig,
@@ -746,6 +747,25 @@ impl Space {
             author,
         )
         .await
+    }
+
+    /// Create a SourceChainWorkspace from this Space
+    pub async fn source_chain_workspace(
+        &self,
+        keystore: MetaLairClient,
+        agent_pubkey: AgentPubKey,
+        dna_def: Arc<DnaDef>,
+    ) -> ConductorResult<SourceChainWorkspace> {
+        Ok(SourceChainWorkspace::new(
+            self.authored_db.clone(),
+            self.dht_db.clone(),
+            self.dht_query_cache.clone(),
+            self.cache_db.clone(),
+            keystore,
+            agent_pubkey,
+            dna_def,
+        )
+        .await?)
     }
 }