+++ crates/holochain/tests/sharded_gossip/mod.rs
@@ -4,14 +4,12 @@ use std::time::Instant;
 use hdk::prelude::*;
 use holo_hash::DhtOpHash;
 use holochain::conductor::config::ConductorConfig;
-use holochain::sweettest::{
-    standard_config, SweetConductor, SweetConductorBatch, SweetDnaFile, SweetInlineZomes,
-};
+use holochain::sweettest::{SweetConductor, SweetConductorBatch, SweetDnaFile, SweetInlineZomes};
 use holochain::test_utils::inline_zomes::{
     batch_create_zome, simple_create_read_zome, simple_crud_zome,
 };
 use holochain::test_utils::network_simulation::{data_zome, generate_test_data};
-use holochain::test_utils::{consistency_10s, consistency_60s, consistency_60s_advanced};
+use holochain::test_utils::{consistency_10s, consistency_60s, consistency_advanced};
 use holochain::{
     conductor::ConductorBuilder, test_utils::consistency::local_machine_session_with_hashes,
 };
@@ -22,7 +20,11 @@ use kitsune_p2p::gossip::sharded_gossip::test_utils::{check_ops_bloom, create_ag
 use kitsune_p2p::KitsuneP2pConfig;
 use kitsune_p2p_types::config::RECENT_THRESHOLD_DEFAULT;
 
-fn make_config(recent: bool, historical: bool, recent_threshold: Option<u64>) -> ConductorConfig {
+fn make_config(
+    recent: bool,
+    historical: bool,
+    recent_threshold: Option<u64>,
+) -> holochain::sweettest::SweetConductorConfig {
     let mut tuning =
         kitsune_p2p_types::config::tuning_params_struct::KitsuneP2pTuningParams::default();
     tuning.gossip_strategy = "sharded-gossip".to_string();
@@ -41,21 +43,12 @@ fn make_config(recent: bool, historical: bool, recent_threshold: Option<u64>) ->
     // so we can fallback to the next one
     tuning.default_rpc_single_timeout_ms = 3_000;
 
-    let mut network = KitsuneP2pConfig::default();
-    network.transport_pool = vec![kitsune_p2p::TransportConfig::Quic {
-        bind_to: None,
-        override_host: None,
-        override_port: None,
-    }];
-    network.tuning_params = Arc::new(tuning);
-    let mut config = standard_config();
-    config.network = Some(network);
-    config
+    holochain::sweettest::SweetConductorConfig::standard().tune(Arc::new(tuning))
 }
 
 #[cfg(feature = "test_utils")]
 #[tokio::test(flavor = "multi_thread")]
-async fn fullsync_sharded_gossip() -> anyhow::Result<()> {
+async fn fullsync_sharded_gossip_low_data() -> anyhow::Result<()> {
     let _g = observability::test_run().ok();
     const NUM_CONDUCTORS: usize = 2;
 
@@ -76,7 +69,7 @@ async fn fullsync_sharded_gossip() -> anyhow::Result<()> {
         .await;
 
     // Wait long enough for Bob to receive gossip
-    consistency_10s([&alice, &bobbo]).await;
+    consistency_60s([&alice, &bobbo]).await;
     // let p2p = conductors[0].envs().p2p().lock().values().next().cloned().unwrap();
     // holochain_state::prelude::dump_tmp(&p2p);
     // holochain_state::prelude::dump_tmp(&alice.env());
@@ -225,7 +218,7 @@ async fn three_way_gossip_historical() {
 /// - 6MB of data can pass from node A to B,
 /// - then A can shut down and C and start up,
 /// - and then that same data passes from B to C.
-async fn three_way_gossip(config: ConductorConfig) {
+async fn three_way_gossip(config: holochain::sweettest::SweetConductorConfig) {
     let mut conductors = SweetConductorBatch::from_config(2, config.clone()).await;
     let start = Instant::now();
 
@@ -292,7 +285,12 @@ async fn three_way_gossip(config: ConductorConfig) {
     conductors.add_conductor(conductor);
     conductors.exchange_peer_info().await;
 
-    consistency_60s_advanced([(&cells[0], false), (&cells[1], true), (&cell, true)]).await;
+    consistency_advanced(
+        [(&cells[0], false), (&cells[1], true), (&cell, true)],
+        30,
+        std::time::Duration::from_secs(1),
+    )
+    .await;
 
     println!(
         "Done waiting for consistency between last two nodes. Elapsed: {:?}",