+++ crates/holochain_integrity_types/src/lib.rs
@@ -263,3 +263,12 @@ impl UnitEnum for () {
         Box::new([].into_iter())
     }
 }
+
+/// A full UnitEnum, or just the unit type of that UnitEnum
+#[derive(Clone, Debug)]
+pub enum UnitEnumEither<E: UnitEnum> {
+    /// The full enum
+    Enum(E),
+    /// Just the unit enum
+    Unit(E::Unit),
+}