+++ nix/modules/devShells.nix
@@ -36,39 +36,108 @@
           ++ (with pkgs; [ cargo-readme cargo-sweep gh gitFull cacert ]);
         };
 
-        coreDev = pkgs.mkShell {
-          inputsFrom = [ self'.devShells.rustDev ];
+        coreDev =
+          let
+            holochainTestDrvs =
+              (
+                lib.attrsets.filterAttrs
+                  (name: package:
+                    # (package.checkPhase or null) != null &&
+                    (builtins.match "^build-holochain-tests.*" name) != null
+                  )
+                  self'.packages
+              );
 
-          packages = with pkgs; [ cargo-nextest ];
+            mkTestScript = name: package:
+              pkgs.writeShellScriptBin (builtins.replaceStrings [ "build-" ] [ "script-" ] name)
+                (
+                  ''
+                    set -xue
+                  ''
+                  # remove the craneLib internals that are part of the checkPhase and
+                  # some characters that would prevent the passing of args
+                  + (
+                    let
+                      cleanCmd = cmd:
+                        (builtins.replaceStrings
+                          [ "cargo --version" "cargoWithProfile" "runHook preCheck" "runHook postCheck" "runHook preBuild" "runHook postBuild" "\n" "\\" "  " ]
+                          [ "" "cargo" "" "" "" "" "" "" " " ]
+                          cmd);
 
-          shellHook = ''
-            export PS1='\n\[\033[1;34m\][coreDev:\w]\$\[\033[0m\] '
+                      checkPhaseClean = cleanCmd (package.checkPhase or "");
+                      buildPhaseClean = cleanCmd (package.buildPhase or "");
+                      checkOrBuildPhase =
+                        if checkPhaseClean != "" then checkPhaseClean
+                        else buildPhaseClean;
+                    in
+                    if checkOrBuildPhase != "" then
+                      ''${checkOrBuildPhase} ''${@}''
+                    else if (package.passthru.dependencies or null) != null then
+                    # recursive call to generate one script call per dependency
+                      builtins.concatStringsSep "\n" (builtins.map (pkg: "${mkTestScript pkg.name pkg}/bin/${pkg.name}") package.passthru.dependencies)
+                    else
+                      throw ''${name} has neither of these
+                              - checkPhase: (${checkPhaseClean})
+                              - buildPhase: (${buildPhaseClean})
+                              - passthru.dependencies
+                            ''
+                  )
+                );
 
-            export HC_TEST_WASM_DIR="$CARGO_TARGET_DIR/.wasm_target"
-            mkdir -p $HC_TEST_WASM_DIR
+          in
+          pkgs.mkShell {
+            inputsFrom = [ self'.devShells.rustDev ] ++ (builtins.attrValues holochainTestDrvs);
 
-            export HC_WASM_CACHE_PATH="$CARGO_TARGET_DIR/.wasm_cache"
-            mkdir -p $HC_WASM_CACHE_PATH
-          '';
-        };
+            packages = with pkgs; [
+              cargo-nextest
 
-        rustDev = pkgs.mkShell
-          {
-            inputsFrom = [
-              self'.packages.holochain
-            ];
+              (pkgs.writeShellScriptBin "script-cargo-regen-lockfiles" ''
+                cargo fetch --locked
+                cargo generate-lockfile --offline --manifest-path=crates/test_utils/wasm/wasm_workspace/Cargo.toml
+                cargo generate-lockfile --offline
+                cargo generate-lockfile --offline --manifest-path=crates/test_utils/wasm/wasm_workspace/Cargo.toml
+              '')
+
+            ]
+
+            # generate one script for each of the "holochain-tests-" prefixed derivations by reusing their checkPhase
+            ++ builtins.attrValues (builtins.mapAttrs mkTestScript holochainTestDrvs)
+            ;
 
             shellHook = ''
-              export CARGO_HOME="$PWD/.cargo"
-              export CARGO_INSTALL_ROOT="$PWD/.cargo"
-              export CARGO_TARGET_DIR="$PWD/target"
-              export CARGO_CACHE_RUSTC_INFO=1
-              export PATH="$CARGO_INSTALL_ROOT/bin:$PATH"
-              export NIX_PATH="nixpkgs=${pkgs.path}"
-            '' + (lib.strings.optionalString pkgs.stdenv.isDarwin ''
-              export DYLD_FALLBACK_LIBRARY_PATH="$(rustc --print sysroot)/lib"
-            '');
+              export PS1='\n\[\033[1;34m\][coreDev:\w]\$\[\033[0m\] '
+
+              export HC_TEST_WASM_DIR="$CARGO_TARGET_DIR/.wasm_target"
+              mkdir -p $HC_TEST_WASM_DIR
+
+              export HC_WASM_CACHE_PATH="$CARGO_TARGET_DIR/.wasm_cache"
+              mkdir -p $HC_WASM_CACHE_PATH
+
+              # Enables the pre-commit hooks
+              ${config.pre-commit.installationScript}
+            '';
           };
+
+        rustDev =
+          pkgs.mkShell
+            {
+              inputsFrom = [
+                self'.packages.holochain
+              ];
+
+              shellHook = ''
+                export CARGO_HOME="$PWD/.cargo"
+                export CARGO_INSTALL_ROOT="$PWD/.cargo"
+                export CARGO_TARGET_DIR="$PWD/target"
+                export CARGO_CACHE_RUSTC_INFO=1
+                export PATH="$CARGO_INSTALL_ROOT/bin:$PATH"
+                export NIX_PATH="nixpkgs=${pkgs.path}"
+              '' + (lib.strings.optionalString pkgs.stdenv.isDarwin ''
+                export DYLD_FALLBACK_LIBRARY_PATH="$(rustc --print sysroot)/lib"
+              '');
+            };
       };
     };
 }
+
+