+++ nix/modules/holochain.nix
@@ -37,37 +37,36 @@
             IOKit
           ]));
 
-        nativeBuildInputs = (with pkgs; [ makeWrapper perl pkg-config ])
+        nativeBuildInputs = (with pkgs; [ makeWrapper perl pkg-config go ])
           ++ lib.optionals pkgs.stdenv.isDarwin
           (with pkgs; [ xcbuild libiconv ]);
       };
 
       # derivation building all dependencies
-      holochainDeps = craneLib.buildDepsOnly (commonArgs // rec {
+      holochainDeps = craneLib.buildDepsOnly (commonArgs // {
         src = flake.config.srcCleanedHolochain;
         doCheck = false;
       });
 
-      holochainDepsRelease = craneLib.buildDepsOnly (commonArgs // rec {
+      holochainDepsRelease = craneLib.buildDepsOnly (commonArgs // {
         CARGO_PROFILE = "release";
         src = flake.config.srcCleanedHolochain;
         doCheck = false;
       });
 
       # derivation with the main crates
-      holochain = (craneLib.buildPackage (commonArgs // {
+      holochain = craneLib.buildPackage (commonArgs // {
         CARGO_PROFILE = "release";
         cargoArtifacts = holochainDepsRelease;
         src = flake.config.srcCleanedHolochain;
         doCheck = false;
-      })) // {
-        src.rev = inputs.holochain.rev;
-      };
+        passthru.src.rev = inputs.holochain.rev;
+      });
 
-      holochainNextestDeps = craneLib.buildDepsOnly (commonArgs // rec {
-        pname = "holochain-nextest";
+      holochainNextestDeps = craneLib.buildDepsOnly (commonArgs // {
+        pname = "holochain-tests-nextest";
         CARGO_PROFILE = "fast-test";
-        nativeBuildInputs = [ pkgs.cargo-nextest ];
+        nativeBuildInputs = commonArgs.nativeBuildInputs ++ [ pkgs.cargo-nextest ];
         buildPhase = ''
           cargo nextest run --no-run \
           ${import ../../.config/test-args.nix} \
@@ -86,6 +85,8 @@
             "conductor::cell::gossip_test::gossip_test"
           ] ++ (lib.optionals (pkgs.system == "x86_64-darwin") [
             "test_reconnect"
+            "timeout::tests::kitsune_backoff"
+            "test_util::switchboard::tests::transitive_peer_gossip"
           ]) ++ (lib.optionals (pkgs.system == "aarch64-darwin") [
             "test_reconnect"
           ]);
@@ -98,6 +99,8 @@
         (commonArgs // {
           __noChroot = pkgs.stdenv.isLinux;
           cargoArtifacts = holochainNextestDeps;
+
+          pname = "holochain-tests-nextest";
 
           preCheck = ''
             export DYLD_FALLBACK_LIBRARY_PATH=$(rustc --print sysroot)/lib
@@ -121,20 +124,9 @@
           '';
         });
 
-      holochain-tests-nextest = craneLib.cargoNextest holochainTestsNextestArgs;
-      holochain-tests-nextest-tx5 = craneLib.cargoNextest
-        (holochainTestsNextestArgs // {
-          pname = "holochain-nextest-tx5";
-          cargoExtraArgs = holochainTestsNextestArgs.cargoExtraArgs + '' \
-            --features tx5 \
-          '';
+      build-holochain-tests-unit = craneLib.cargoNextest holochainTestsNextestArgs;
 
-          nativeBuildInputs = holochainTestsNextestArgs.nativeBuildInputs ++ [
-            pkgs.go
-          ];
-        });
-
-      holochain-tests-fmt = craneLib.cargoFmt (commonArgs // {
+      build-holochain-tests-static-fmt = craneLib.cargoFmt (commonArgs // {
         src = flake.config.srcCleanedHolochain;
         cargoArtifacts = null;
         doCheck = false;
@@ -143,18 +135,13 @@
         dontFixup = true;
       });
 
-      holochain-tests-clippy = craneLib.cargoClippy (commonArgs // {
+      build-holochain-tests-static-clippy = craneLib.cargoClippy (commonArgs // {
         pname = "holochain-tests-clippy";
         src = flake.config.srcCleanedHolochain;
         cargoArtifacts = holochainDeps;
         doCheck = false;
 
-        cargoClippyExtraArgs = ''
-          -- \
-          -A clippy::nursery -D clippy::style -A clippy::cargo \
-          -A clippy::pedantic -A clippy::restriction \
-          -D clippy::complexity -D clippy::perf -D clippy::correctness
-        '';
+        cargoClippyExtraArgs = "-- ${import ../../.config/clippy-args.nix}";
 
         dontPatchELF = true;
         dontFixup = true;
@@ -167,23 +154,22 @@
 
       holochainWasmArgs = (commonArgs // {
         pname = "holochain-tests-wasm";
-        cargoExtraArgs =
-          "--lib --all-features";
 
-        cargoToml = "${self}/crates/test_utils/wasm/wasm_workspace/Cargo.toml";
-        cargoLock = "${self}/crates/test_utils/wasm/wasm_workspace/Cargo.lock";
-
-        postUnpack = ''
-          cd $sourceRoot/crates/test_utils/wasm/wasm_workspace
-          sourceRoot="."
+        postConfigure = ''
+          export CARGO_TARGET_DIR=''${CARGO_TARGET_DIR:-$PWD/target}
         '';
+
+        cargoExtraArgs =
+          "--lib --all-features --manifest-path=crates/test_utils/wasm/wasm_workspace/Cargo.toml";
+
+        cargoLock = "${flake.config.srcCleanedHolochain}/crates/test_utils/wasm/wasm_workspace/Cargo.lock";
       });
 
       holochainDepsWasm = craneLib.buildDepsOnly (holochainWasmArgs // {
         cargoArtifacts = null;
       });
 
-      holochain-tests-wasm = craneLib.cargoTest (holochainWasmArgs // {
+      build-holochain-tests-unit-wasm = craneLib.cargoTest (holochainWasmArgs // {
         cargoArtifacts = holochainDepsWasm;
 
         dontPatchELF = true;
@@ -195,17 +181,48 @@
         '';
       });
 
-      holochain-tests-doc = craneLib.cargoDoc (commonArgs // {
+      build-holochain-tests-static-doc = craneLib.cargoDoc (commonArgs // {
         pname = "holochain-tests-docs";
         cargoArtifacts = holochainDeps;
       });
 
+
+
+      # meta packages to build multiple test packages at once
+      build-holochain-tests-unit-all = config.lib.mkMetaPkg "holochain-tests-unit-all" [
+        build-holochain-tests-unit
+        build-holochain-tests-unit-wasm
+      ];
+
+      build-holochain-tests-static-all = config.lib.mkMetaPkg "holochain-tests-static-all" [
+        build-holochain-tests-static-doc
+        build-holochain-tests-static-fmt
+        build-holochain-tests-static-clippy
+      ];
+
+      build-holochain-tests-all = config.lib.mkMetaPkg "build-holochain-tests-all" [
+        build-holochain-tests-unit-all
+        build-holochain-tests-static-all
+      ];
+
     in
     {
-      packages = {
-        inherit holochain holochain-tests-nextest holochain-tests-nextest-tx5 holochain-tests-doc;
+      packages =
+        {
+          inherit
+            holochain
 
-        inherit holochain-tests-wasm holochain-tests-fmt holochain-tests-clippy;
-      };
+            build-holochain-tests-unit
+            build-holochain-tests-unit-wasm
+            build-holochain-tests-unit-all
+
+            build-holochain-tests-static-doc
+            build-holochain-tests-static-fmt
+            build-holochain-tests-static-clippy
+            build-holochain-tests-static-all
+
+            build-holochain-tests-all
+            ;
+        };
     };
 }